python_backend/api/database.py
import mysql.connector
import os
from dotenv import load_dotenv

# Load environment variables from .env file
# The path is relative to the project root where you run `uvicorn`
load_dotenv()

def get_db_connection():
    """Establishes a connection to the MySQL database."""
    try:
        conn = mysql.connector.connect(
            host=os.getenv("DB_HOST"),
            user=os.getenv("DB_USER"),
            password=os.getenv("DB_PASSWORD"),
            database=os.getenv("DB_NAME")
        )
        return conn
    except mysql.connector.Error as err:
        print(f"Error connecting to database: {err}")
        return None 