python_backend/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv 